Add a ClientDispatcher, reduce complexity of McClient, switch to FluentResults.
Used in conjunction with the SessionDispatcher, is used to maintain separation of concerns inside McClient.

Due to the complexity of preceding code, many things are still broken and/or requiring implementation detail and testing.

FluentResults is added as an alternative to passing values instead of through `ref` parameters, to accomodate for async/await syntax.