replace static settings with instanced settings
All debug printing is delegated to Serilog instead

Some constants are moved into a dedicated constants file (Const.cs)

All code that references Settings now require an extra parameter that takes an instance of Settings.

None of the code here is written in an ideal way- there are a lot of passing around of the Settings class and the ProxyHandler class due to the way the underlying code works.